+++ AutomatedLab/AutomatedLabVirtualMachines.psm1
@@ -1836,12 +1836,14 @@ function Get-LabVM
 }
 #endregion Get-LabVM
 
-#region Set-LabAutoLogon
-function Set-LabAutoLogon
+#region Enable-LabAutoLogon
+function Enable-LabAutoLogon
 {
+    [CmdletBinding()]
+    [Alias('Set-LabAutoLogon')]
     param
     (
-        [Parameter(Mandatory)]
+        [Parameter()]
         [string[]]
         $ComputerName
     )
@@ -1875,13 +1877,37 @@ function Set-LabAutoLogon
         } -Variable (Get-Variable InvokeParameters) -NoDisplay
     }
 }
-#endregion Set-LabAutoLogon
+#endregion
+
+#region Disable-LabAutoLogon
+function Disable-LabAutoLogon
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter()]
+        [string[]]
+        $ComputerName
+    )
+
+    Write-Verbose -Message "Disabling autologon on $($ComputerName.Count) machines"
+
+    $Machines = Get-LabVm @PSBoundParameters
+
+    Invoke-LabCommand -ActivityName "Disabling AutoLogon on $($ComputerName.Count) machines" -ComputerName $Machines -ScriptBlock {
+        Set-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon" -Name AutoAdminLogon -Value 0 -Type String -Force
+        Remove-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon" -Name DefaultPassword -Force -ErrorAction SilentlyContinue
+    } -NoDisplay
+}
+#endregion
 
 #region Test-LabAutoLogon
 function Test-LabAutoLogon
 {
+    [CmdletBinding()]
     param
     (
+        [Parameter()]
         [string[]]
         $ComputerName
     )
@@ -2178,3 +2204,65 @@ function Remove-LabVMSnapshot
     Write-LogFunctionExit
 }
 #endregion Remove-LabVMSnapshot
+
+#region Get-LabVmSnapshot
+function Get-LabVMSnapshot
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter()]
+        [string[]]
+        $ComputerName,
+
+        [Parameter()]
+        [string]
+        $SnapshotName
+    )
+
+    Write-LogFunctionEntry
+
+    if (-not (Get-LabVM))
+    {
+        Write-Error 'No machine definitions imported, so there is nothing to do. Please use Import-Lab first'
+        return
+    }
+
+    $lab = Get-Lab
+
+    if ($ComputerName)
+    {
+        $machines = Get-LabVM -IncludeLinux | Where-Object -Property Name -in $ComputerName
+    }
+    else
+    {
+        $machines = Get-LabVm -IncludeLinux
+    }
+
+    if (-not $machines)
+    {
+        $message = 'No machine found to remove the snapshot. Either the given name is wrong or there is no machine defined yet'
+        Write-LogFunctionExitWithError -Message $message
+        return
+    }
+
+    $parameters = @{
+        VMName = $machines
+        ErrorAction = 'SilentlyContinue'
+    }
+    
+    if ($SnapshotName)
+    {
+        $parameters.Name = $SnapshotName
+    }
+
+    switch ($lab.DefaultVirtualizationEngine)
+    {
+        'HyperV' { Get-LWHypervVMSnapshot @parameters}
+        'Azure'  { Get-LWAzureVmSnapshot @parameters}
+        'VMWare' { Write-ScreenInfo -Type Warning -Message 'No VMWare snapshots possible, nothing will be listed'}
+    }
+
+    Write-LogFunctionExit
+}
+#endregion