+++ LabXml/Machines/OperatingSystem.cs
@@ -25,6 +25,8 @@ namespace AutomatedLab
             {"2012-R2-Datacenter", "Windows Server 2012 R2 Datacenter (Server with a GUI)" },
             {"2016-Datacenter", "Windows Server 2016 Datacenter (Desktop Experience)" },
             {"2016-Datacenter-Server-Core", "Windows Server 2016 Datacenter" },
+            {"2019-Datacenter", "Windows Server 2019 Datacenter (Desktop Experience)" },
+            {"2019-Datacenter-Core", "Windows Server 2019 Datacenter" },
             {"Datacenter-Core-1803-with-Containers-smalldisk", "Windows Server Datacenter" },
             {"Win81-Ent-N-x64", "Windows 8.1 Enterprise" },
             {"Windows-10-N-x64", "Windows 10 Enterprise" },